Refactor and modernize Feather-Trace
Major rewrite of the codebase. Benefits:
- Added support for x86_64. The trigger code was contributed by Andrea Bastoni.
- Cleaner separation of the userspace code for dynamic libraries and
  static compilation.
- Recent versions of GNU ld only PROVIDE() the start and end address of
  orphan sections, but we need these symbols even if they are not
  referenced by the compilation unit. We now use a linker script to
  generate needed symbols in dynamic libraries.
- Removed the need for INIT_FT_EVENTS() by using a constructor function, i.e.,
  marked the init function with __attribute__((constructor)).
- Added __attribute__((used)) to Feather-Trace callbacks to suppress gcc warnings.
- Arguments to trigger macros are now automatically cast to long (which is
  required anyway) to suppress confusing error message in case the argument is
  an int on x86_64.

+++ include/feather_buffer.h
@@ -67,7 +67,7 @@ static inline int init_ft_buffer(struct ft_buffer*	buf,
 				 char*			slots,
 				 void* 			buffer_mem) 
 {
-	int i = 0;
+	unsigned int i = 0;
 	if (!slot_count || UINT_MAX % slot_count != slot_count - 1) {
 		/* The slot count must divide UNIT_MAX + 1 so that when it
 		 * wraps around the index correctly points to 0.
@@ -118,7 +118,7 @@ static inline void ft_buffer_finish_write(struct ft_buffer* buf, void *ptr)
 static inline int ft_buffer_read(struct ft_buffer* buf, void* dest)
 {
 	unsigned int idx;
-	if (buf->free_count == buf->slot_count) 
+	if (buf->free_count == (int) buf->slot_count) 
 		/* nothing available */
 		return 0;
 	idx = buf->read_idx % buf->slot_count;
